backend/requirements.txt
flask
flask-cors
pymongo
python-dotenv
passlib[bcrypt]
pyjwt
reportlab
openai
